# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MDGA Enterprise B2B SaaS",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/clear_db": {
      "delete": {
        "tags": [
          "Admin v1"
        ],
        "summary": "Clear Db",
        "operationId": "clear_db_api_v1_clear_db_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reset_schema": {
      "post": {
        "tags": [
          "Admin v1"
        ],
        "summary": "Reset Schema",
        "operationId": "reset_schema_api_v1_reset_schema_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/inject": {
      "post": {
        "tags": [
          "Admin v1"
        ],
        "summary": "Demo Inject",
        "operationId": "demo_inject_api_v1_demo_inject_post",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulate/governance": {
      "post": {
        "tags": [
          "Admin v1"
        ],
        "summary": "Simulate Governance",
        "operationId": "simulate_governance_api_v1_simulate_governance_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_simulate_governance_api_v1_simulate_governance_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ax-data/yield-prediction": {
      "get": {
        "tags": [
          "AX Synthetic Data v1"
        ],
        "summary": "Get Yield Prediction",
        "operationId": "get_yield_prediction_api_v1_ax_data_yield_prediction_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target region (e.g. \ub300\uad6c\uad11\uc5ed\uc2dc \ubd81\uad6c)",
              "title": "Region"
            },
            "description": "Target region (e.g. \ub300\uad6c\uad11\uc5ed\uc2dc \ubd81\uad6c)"
          },
          {
            "name": "crop",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target crop (e.g. \uc0ac\uacfc)",
              "title": "Crop"
            },
            "description": "Target crop (e.g. \uc0ac\uacfc)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ax-data/crop-simulator": {
      "get": {
        "tags": [
          "AX Synthetic Data v1"
        ],
        "summary": "Get Crop Simulator",
        "operationId": "get_crop_simulator_api_v1_ax_data_crop_simulator_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target region",
              "title": "Region"
            },
            "description": "Target region"
          },
          {
            "name": "crop",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target crop",
              "title": "Crop"
            },
            "description": "Target crop"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ax-data/oversupply-risk": {
      "get": {
        "tags": [
          "AX Synthetic Data v1"
        ],
        "summary": "Get Oversupply Risk",
        "operationId": "get_oversupply_risk_api_v1_ax_data_oversupply_risk_get",
        "parameters": [
          {
            "name": "crop",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target crop",
              "title": "Crop"
            },
            "description": "Target crop"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ax-data/resource-efficiency": {
      "get": {
        "tags": [
          "AX Synthetic Data v1"
        ],
        "summary": "Get Resource Efficiency",
        "operationId": "get_resource_efficiency_api_v1_ax_data_resource_efficiency_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target region",
              "title": "Region"
            },
            "description": "Target region"
          },
          {
            "name": "crop",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target crop",
              "title": "Crop"
            },
            "description": "Target crop"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ax-data/history": {
      "get": {
        "tags": [
          "AX Synthetic Data v1"
        ],
        "summary": "Get Synthetic Data History",
        "description": "Fetch cached synthetic data from the DB",
        "operationId": "get_synthetic_data_history_api_v1_ax_data_history_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Region"
            }
          },
          {
            "name": "data_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Data Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_simulate_governance_api_v1_simulate_governance_post": {
        "properties": {
          "budget": {
            "type": "integer",
            "title": "Budget"
          },
          "region": {
            "type": "string",
            "title": "Region"
          }
        },
        "type": "object",
        "required": [
          "budget",
          "region"
        ],
        "title": "Body_simulate_governance_api_v1_simulate_governance_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}